+++ crates/aero-usb/tests/xhci_mmio_smoke.rs
@@ -3,7 +3,7 @@ use aero_usb::xhci::XhciController;
 
 fn op_base(ctrl: &mut XhciController) -> u64 {
     let cap0 = ctrl.mmio_read(regs::cap::CAPLENGTH as u64, 4);
-    (cap0 & 0xff) as u64
+    cap0 & 0xff
 }
 
 #[test]
